requirements.txt
altair==5.5.0
attrs==25.3.0
Automat==24.8.1
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
contourpy==1.3.0
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
defusedxml==0.7.1
dotenv==0.9.9
entrypoints==0.4
Faker==37.1.0
favicon==0.7.0
filelock==3.18.0
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
htbuilder==0.9.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lxml==5.3.1
Markdown==3.7
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.9.4
narwhals==1.33.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
parsel==1.10.0
pillow==11.1.0
plotly==6.0.1
prometheus_client==0.21.1
Protego==0.4.0
protobuf==5.29.4
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydeck==0.9.1
PyDispatcher==2.0.7
pymdown-extensions==10.14.3
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
queuelib==1.8.0
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rpds-py==0.24.0
Scrapy==2.12.0
seaborn==0.13.2
service-identity==24.2.0
six==1.17.0
smmap==5.0.2
soupsieve==2.6
SQLAlchemy==2.0.40
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.44.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==9.1.2
tldextract==5.1.3
toml==0.10.2
tornado==6.4.2
Twisted==24.11.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uuid==1.30
validators==0.34.0
w3lib==2.3.1
watchdog==6.0.0
zope.interface==7.2
